restructuring, added sourcedoc
